Fix tests and move manager to schedule package
